requirements.txt
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
